Improves flexibility of the bdpar log system
Added the possibility of storing the log in a file. Moreover, the different levels of the messages allow the output to be filtered. One of the utilities that have been implemented with the log, is to show the content of the instances in DEBUG mode to see its status between tasks.