Implements parallel mode
The functionality to parallelize the execution of the instances has been implemented. In addition, the log has been improved to be compatible with thread handling, using the "global options". It is important to note that the console log is disabled in the child threads (the file log must be used).

Additionally, the "Collate" field has been added to the "DESCRIPTION" file to better control the order in which the package files are loaded.